refactor: OrderProductItem 위치 변경
이름 바꿀까 고민했는데 그냥 하기로 함
