+++ src/main/java/com/fourweekdays/fourweekdays/outbound/model/entity/OutboundProductItem.java
@@ -1,6 +1,6 @@
 package com.fourweekdays.fourweekdays.outbound.model.entity;
 
-import com.fourweekdays.fourweekdays.order.OrderProductItem;
+import com.fourweekdays.fourweekdays.order.model.entity.OrderProductItem;
 import com.fourweekdays.fourweekdays.product.model.entity.Product;
 import jakarta.persistence.*;
 import lombok.*;
